requirements.txt
# PIPE Domain Bot System - Python Dependencies

# Core dependencies
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Async support
asyncio>=3.4.3

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1

# Code quality
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
pylint>=2.17.5

# Development
ipython>=8.14.0
